+++ dart/geobase/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.5.0
+
+NOTE: Version 0.5.0 currently under development (0.5.0-dev.0).
+
+🐛 Bugs fixed:
+* [Clipping when projection WGS 84 to Web Mercator metric coordinates #123](https://github.com/navibyte/geospatial/issues/123)
+
 ## 0.4.2
 
 📚 Documentation updates: